+++ module/strategy/StrategiseLook/src/StrategiseLook.cpp
@@ -40,7 +40,7 @@ namespace module::strategy {
             [this](const FilteredBall& ball) {
                 // If we have a ball and it is recent, look at it
                 if (NUClear::clock::now() - ball.time_of_measurement < cfg.ball_search_timeout) {
-                    emit<Task>(std::make_unique<Look>(ball.rBCt.cast<double>(), true));
+                    emit<Task>(std::make_unique<Look>(ball.rBCt, true));
                 }
             });
 
@@ -51,7 +51,7 @@ namespace module::strategy {
                 // If we have goals, with at least one measurement and the goals are recent, look at the goals
                 if (!goals.goals.empty() && (NUClear::clock::now() - goals.timestamp < cfg.goal_search_timeout)) {
                     // Convert goal measurement to cartesian coordinates
-                    Eigen::Vector3d rGCc = sphericalToCartesian(goals.goals[0].measurements[0].srGCc.cast<double>());
+                    Eigen::Vector3d rGCc = sphericalToCartesian(goals.goals[0].measurements[0].srGCc);
                     // Convert to torso space
                     Eigen::Vector3d rGCt = Eigen::Isometry3d(sensors.Htw * goals.Hcw.inverse()).rotation() * rGCc;
                     // Look at the goal
